requirements/base.txt
click
matplotlib
zipp
pytorch
pytorch-lightning
tabulate
torchvision
tensorboard
cml-pam